requirements.txt
requests
telethon
python-dotenv
asyncio
